#3686 comment: refactoring: adjusted path to common folder in imports
