+++ api/endpoints/common/entity_base.py
