+++ api/endpoints/common/sql_queries.py
