+++ api/endpoints/common/sql_templates.py
