+++ api/local/admin_tasks/general_admin/anon_project_specific_user_id_to_user_info.py
@@ -20,15 +20,6 @@ This script parses a CSV file containing anon_project_specific_user_ids and outp
 including name, email and id
 """
 
-import csv
-import os
-from tkinter import Tk
-from tkinter.filedialog import askopenfilename
-
-import api.endpoints.user as u
-import api.common.utilities as utils
-from api.endpoints.user_group import UserGroup
-from api.endpoints.user_group_membership import UserGroupMembership
 from api.local.admin_tasks.admin_tasks_utilities import CsvImporter
