+++ api/local/admin_tasks/hubspot_admin/create_thiscovery_contact_properties.py
@@ -1,5 +1,5 @@
 from http import HTTPStatus
-from api.common.hubspot import HubSpotClient
+from api.endpoints.common.hubspot import HubSpotClient
 from thiscovery_lib.utilities import DetailedValueError
