+++ api/local/admin_tasks/hubspot_admin/create_timeline_event_type_for_task_signup.py
@@ -1,5 +1,5 @@
 import thiscovery_lib.dynamodb_utilities as ddb
-import api.common.hubspot as hs
+import api.endpoints.common.hubspot as hs
 from thiscovery_lib.utilities import get_aws_namespace
