+++ api/local/admin_tasks/hubspot_admin/renew_hubspot_token.py
@@ -1,4 +1,4 @@
-from api.common.hubspot import HubSpotClient, SingleSendClient
+from api.endpoints.common.hubspot import HubSpotClient, SingleSendClient
 
 
 def renew_token(client_class):
