+++ api/local/admin_tasks/task_setup/output_anon_project_specific_user_ids_for_test_group.py
@@ -36,8 +36,8 @@ import csv
 
 import api.endpoints.project as p
 import api.endpoints.user_project as up
-import api.common.pg_utilities as pg_utils
-import api.common.sql_queries as sq
+import api.endpoints.common.pg_utilities as pg_utils
+import api.endpoints.common.sql_queries as sq
 import thiscovery_lib.utilities as utils
