+++ api/tests/test_scripts/integration_tests/test_admin_tasks.py
@@ -15,16 +15,17 @@
 #   A copy of the GNU Affero General Public License is available in the
 #   docs folder of this project.  It is also available www.gnu.org/licenses/
 #
+import testing_utilities as test_utils
 import os
 from pprint import pprint
 
-import api.common.pg_utilities as pg_utils
-import api.common.sql_queries as sql_q
+import api.endpoints.common.pg_utilities as pg_utils
+import api.endpoints.common.sql_queries as sql_q
 import thiscovery_lib.utilities as utils
-import testing_utilities as test_utils
+
 
 from thiscovery_lib.dynamodb_utilities import Dynamodb
-from api.common.dev_config import QUALTRICS_TEST_OBJECTS
+from api.endpoints.common.dev_config import QUALTRICS_TEST_OBJECTS
 from api.local.admin_tasks.qualtrics_admin.convert_responses_to_contact_list_input_format import ResponsesToContactListConverter
 from api.local.admin_tasks.qualtrics_admin.republish_survey_with_user_specific_links import DistributionLinksGenerator
 from api.local.admin_tasks.task_setup.output_anon_project_specific_user_ids_for_test_group import ProcessManager as OutputAnonIds
