+++ api/tests/test_scripts/unit_tests/test_db_views.py
@@ -16,7 +16,7 @@
 #   docs folder of this project.  It is also available www.gnu.org/licenses/
 #
 import testing_utilities as test_utils  # this should be the first import; it sets env variables
-import common.pg_utilities as pg_utils
+import api.endpoints.common.pg_utilities as pg_utils
 
 
 class DbViewsTestCase(test_utils.DbTestCase):
