+++ api/tests/test_scripts/unit_tests/test_notifications.py
@@ -24,14 +24,14 @@ from http import HTTPStatus
 from thiscovery_lib.dynamodb_utilities import Dynamodb
 
 import api.endpoints.notification_process as np
-import common.notifications as notific
-import common.notification_send as notific_send
+import api.endpoints.common.notifications as notific
+import api.endpoints.common.notification_send as notific_send
 import thiscovery_lib.utilities as utils
 
-from common.hubspot import HubSpotClient
-from common.notifications import NotificationStatus, NotificationAttributes, NotificationType, delete_all_notifications, get_notifications, \
+from api.endpoints.common.hubspot import HubSpotClient
+from api.endpoints.common.notifications import NotificationStatus, NotificationAttributes, NotificationType, delete_all_notifications, get_notifications, \
     mark_notification_failure
-from common.notification_send import notify_new_user_registration, notify_new_task_signup, notify_user_login, new_transactional_email_notification
+from api.endpoints.common.notification_send import notify_new_user_registration, notify_new_task_signup, notify_user_login, new_transactional_email_notification
 from thiscovery_lib.utilities import get_country_name, DetailedValueError
 from test_transactional_email import test_email_dict
